Use multi-stage build in Dockerfile
This was a bit of a learning experience, but really obvious in retrospect.  The needs for building are different from the needs for running.  Using a multistage build, I can run the linter and do all of my building but not have to include all of the detritus and unnecessary source files in my final distribution.