Stop fighting Jest's setup
I must have missed something in the quick setup guides for Jest, but the ones that I saw skipped right over where the tests are put *inline* with the source files.  I'm more familiar with patterns where tests are separated from the code.  While working on some pathing stuff, I realized that the default patterns capture tests that are adjacent to their source files.  This simplifies a lot, but now I need to exclude tests from the final build.