Explore a builder pattern for BlogPosts
Okay.  Wow.  I don't know if I want to hold on to this one.  But holy crap, allocations!

Here's the goal: I wanted to explore a very type-safe builder pattern that forced me to make sure that every single property for a BlogPost was set.  Here's the basic scenario:

* I wanted to avoid the potential for callers to mix up `title` and `slug`.  They're the same time and adjacent and it'd be an easy mistake to make.
* I didn't want to use object deconstruction as a parameters list.  It's nice, but I didn't want to have to add runtime asserts or anything.  Let's catch issues at compile time.
* This *feels* like something that our tooling can handle for us.

I think I found a solution in Typescript.  I'm not keen on the API between the builder and the object, but it's fine for now.  My bigger concern is how this scales at runtime in terms of object allocations.  This feels really inefficient to make two allocations for every setter.  It may not be a concern in practice, but the general inefficiency irks me.  I'll live with it for now.